chat_project/requirements.txt
# requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15          # ← NEW
psycopg2-binary>=2.9
django-environ>=0.9
gunicorn>=21.2
django-storages>=1.14
boto3>=1.34
Pillow>=10.0
python-magic>=0.4.27               # optional but recommended with libmagic-dev
